+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change Log
+10-3-20
+
+  * Added EOC (end of cycle) to NoteSeq16
+  * Added right click options to simple clock
+  * Added tooltips to all knobs and buttons
+  * Can now enter a BPM in simple clock and see it in the tooltip
+  * Added module manual urls
 
 9-1-20
   * Added EOC (end of cycle) to NoteSeq